monolit/requirements/requirements.txt
Django==2.2
django-appconf==1.0.3
django-ckeditor==5.6.1
django-cleanup==3.2.0
django-environ==0.4.5
django-imagekit==4.0.2
django-js-asset==1.2.2
django-location-field==2.1.0
django-unused-media==0.1.13
pilkit==2.0
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
